Added more RigidBody unit test
- Fixed minor errors in the RigidBody class.
- Changed the RigidBody class interface so now every attrbiute is access
  only with getters and setters.
- Added force unit tests.
- Fixed RigidBody update methods.
- Fixed style inconsistencies in the Logger, physics system, collision
  system and tests.
- Changed some size_t to ints.

+++ faze-engine/include/fe/collision/SupportPoint.h
@@ -57,7 +57,7 @@ namespace fe { namespace collision {
 		 *
 		 * @param	i the first or the second collider given in the constructor
 		 * @return	the coordinates of the SupportPoint in local space */
-		inline glm::vec3 getLocalPosition(size_t i) const
+		inline glm::vec3 getLocalPosition(int i) const
 		{ return mLocalPosition[i]; };
 
 		/** Returns the coordinates in world space of the SupportPoint relative
@@ -65,7 +65,7 @@ namespace fe { namespace collision {
 		 *
 		 * @param	i the first or the second collider given in the constructor
 		 * @return	the coordinates of the SupportPoint in world space */
-		inline glm::vec3 getWorldPosition(size_t i) const
+		inline glm::vec3 getWorldPosition(int i) const
 		{ return mWorldPosition[i]; };
 	};
